+++ libs/designsystem/src/lib/components/list/directives/infinite-scroll.directive.ts
@@ -12,7 +12,7 @@ import {
 import { Subject, fromEvent } from 'rxjs';
 import { debounceTime, takeUntil, filter, map } from 'rxjs/operators';
 
-import { WindowRef } from '../../../helpers/window-ref';
+import { WindowRef } from '../../../types/window-ref';
 
 import { Scroll } from './scroll.model';
