+++ libs/designsystem/src/lib/components/modal/modal-wrapper/compact/modal-compact-wrapper.component.ts
@@ -12,7 +12,7 @@ import { Subject } from 'rxjs';
 import { ModalConfig } from '../config/modal-config';
 import { COMPONENT_PROPS } from '../config/modal-config.helper';
 import { Modal } from '../../services/modal.interfaces';
-import { WindowRef } from '../../../../helpers/window-ref';
+import { WindowRef } from '../../../../types/window-ref';
 
 @Component({
   selector: 'kirby-modal-compact-wrapper',
