+++ libs/designsystem/src/lib/components/modal/services/modal.helper.spec.ts
@@ -3,7 +3,7 @@ import { IonicModule, ModalController as IonicModalController } from '@ionic/ang
 import { createService } from '@ngneat/spectator';
 
 import { DesignTokenHelper } from '../../../helpers/design-token-helper';
-import { WindowRef } from '../../../helpers/window-ref';
+import { WindowRef } from '../../../types/window-ref';
 import { TestHelper } from '../../../testing/test-helper';
 import { ModalHelper } from './modal.helper';
 import { Overlay, Modal } from './modal.interfaces';
