+++ libs/designsystem/src/lib/helpers/index.ts
@@ -1,4 +1,3 @@
 export * from './color-helper';
 export * from './design-token-helper';
 export * from './theme-color.type';
-export * from './window-ref';
